requirements.txt
flask==2.3.2
transformers==4.30.2
torch==2.0.1
sentencepiece==0.1.99
google-generativeai==0.3.0
python-dotenv==1.0.0
scikit-learn==1.2.2
pandas==1.5.3
numpy==1.24.3
accelerate>=0.20.1
transformers[torch]
matplotlib
seaborn